pipeline/requirements.txt
anthropic>=0.42.0
PyMuPDF>=1.25.0
weasyprint>=63.0
python-docx>=1.1.0
